test(tests-integration): add setup and state reader spawn to e2e integration test
commit-id:c4f66063
